made Header nav links and Navbar navlinks into 2 separate components
